+++ rom.py
@@ -253,6 +253,7 @@ def generate_output(world: "PokemonPlatinumWorld", output_directory: str, patch:
     add_opt_byte("blind_trainers")
     add_opt_byte("fps60")
     add_opt_byte("hm_cut_ins")
+    add_opt_byte("buck_pos")
 
     if len(ap_bin) % 2 == 1:
         ap_bin += b'\x00'